Time: 28 ms (34.31%), Space: 45.8 MB (65.33%) - LeetHub

+++ 2976-minimum-cost-to-convert-string-i/2976-minimum-cost-to-convert-string-i.java
@@ -19,7 +19,7 @@ class Solution {
             }
         }
     }
-    int totalcost=0;
+    long totalcost=0;
     for(int i=0;i<source.length();i++){
         if(source.charAt(i)==target.charAt(i)){
             continue;
